Move Residency Reducer
close DCOS-14728
