+++ plugins/services/src/js/reducers/JSONSingleContainerReducers.js
@@ -5,7 +5,7 @@ import {JSONReducer as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONReducer as healthChecks} from './serviceForm/HealthChecks';
 import {JSONReducer as labels} from './serviceForm/JSONReducers/Labels';
 import {JSONReducer as portDefinitions} from './serviceForm/PortDefinitions';
-import {JSONReducer as residency} from './serviceForm/Residency';
+import {JSONReducer as residency} from './serviceForm/JSONReducers/Residency';
 import {JSONReducer as ipAddress} from './serviceForm/JSONReducers/IpAddress';
 import {JSONReducer as requirePorts} from './serviceForm/RequirePorts';
 import {
